Refactor meet validation to meet_validation.py and fix rules limits & relay event counts
